third_party/blink/web_tests/editing/selection/caret-after-keypress-expected.txt

This tests that 'input' event listener gets correct caret position after keypress.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test.value is ""
PASS test.selectionStart is 0
PASS test.selectionEnd is 0


PASS test.value is "a"
PASS test.selectionStart is 1
PASS test.selectionEnd is 1


PASS test.value is "ab"
PASS test.selectionStart is 2
PASS test.selectionEnd is 2


PASS test.value is "abc"
PASS test.selectionStart is 3
PASS test.selectionEnd is 3


PASS test.value is "ab"
PASS test.selectionStart is 2
PASS test.selectionEnd is 2


PASS test.value is "a"
PASS test.selectionStart is 1
PASS test.selectionEnd is 1


PASS test.value is ""
PASS test.selectionStart is 0
PASS test.selectionEnd is 0


PASS successfullyParsed is true

TEST COMPLETE
